requirements.txt

pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
umap-learn>=0.5.0
Pillow>=9.0.0
python-igraph>=0.10.0
leidenalg>=0.10.0
scanpy>=1.9.0
scipy>=1.9.0
pathlib
argparse
logging
json5
datetime



